Update patch set 18
Patch Set 18:

The way the Exynos repositories are laid out, most of the device-specific items should be moved to ExynosHWCModule.h if possible. With some more indirection for the existing S3C definitions, it should be possible to move most of the decon header specific items to ExynosHWCModule.h.

Patch-set: 18