Update patch set 21
Patch Set 21: Code-Review-1

Something is not right here.
A random 'still image' (looks like a wallpaper, no graphical glitches or anything) is overlaid on top of the 'real image' our display should be drawing.
You can see the real image when quickly pressing the power button.

Also, Exynos7420 does not expose gscaler nodes anymore, so the support for that should be removed for the recent DECON framebuffer generations.

Patch-set: 21